+++ src/common/border/border.h
@@ -11,6 +11,7 @@ struct border_window
 border_window *CreateBorderWindow(int X, int Y, int W, int H, int BorderWidth, int BorderRadius, unsigned int BorderColor);
 void UpdateBorderWindowRect(border_window *Border, int X, int Y, int W, int H);
 void UpdateBorderWindowColor(border_window *Border, unsigned Color);
+void UpdateBorderWindowWidth(border_window *Border, int BorderWidth);
 void DestroyBorderWindow(border_window *Border);
 
 #endif