+++ src/plugins/border/CHANGELOG.md
@@ -1,3 +1,11 @@
+### HEAD -  not yet released
+
+#### added runtime command
+
+ - command to change border width during runtime: `chunkc border::width <number>`
+
+----------
+
 ### version 0.3.0
 
 #### other changes